Sorting algorithm applies to quests
- Separated default asset sorting logic, applied to quest cards as well
